RI2024G05-335
Tests for homepage updated and code changed according to feedback of the code review

+++ app/src/main/java/com/example/riseandroid/LumiereApplication.kt
@@ -23,7 +23,7 @@ class LumiereApplication : Application() {
         container = DefaultAppContainer(context = applicationContext)
     }
 
-    fun scheduleNotification(context: Context, movieId: Long, movieName: String, location: String, date: String, dateTime: Calendar) {
+    fun scheduleNotification(context: Context, movieId: Int, movieName: String, location: String, date: String, dateTime: Calendar) {
         val intent = Intent(context, NotificationReceiver::class.java).apply {
             putExtra("MOVIE_ID", movieId)
             putExtra("MOVIE_NAME", movieName)
@@ -43,7 +43,7 @@ class LumiereApplication : Application() {
 
 
 
-    fun displayImmediateNotification(context: Context, movieId: Long, movieName: String, location: String, date: String) {
+    fun displayImmediateNotification(context: Context, movieId: Int, movieName: String, location: String, date: String) {
         if (NotificationManagerCompat.from(context).areNotificationsEnabled()) {
             val notification = NotificationCompat.Builder(context, "REMINDER_CHANNEL")
                 .setSmallIcon(R.drawable.ic_launcher_foreground) // Replace with your app's icon
